day11 both part completed

+++ day11/advent-of-code-day11/src/main/java/Pos.java
@@ -1,21 +1,21 @@
 public class Pos {
-    private final int posX;
-    private final int posY;
+    private final long posX;
+    private final long posY;
 
     public Pos(int posX, int posY) {
         this.posX = posX;
         this.posY = posY;
     }
 
-    public int getPosX() {
+    public long getPosX() {
         return posX;
     }
 
-    public int getPosY() {
+    public long getPosY() {
         return posY;
     }
 
-    public int calcDistance(Pos otherPos) {
+    public long calcDistance(Pos otherPos) {
         return Math.abs(this.posX - otherPos.getPosX()) + Math.abs(this.posY - otherPos.getPosY());
     }
 }
